refactor API to use more semantic primitives
make API stop using mainly float64s where integers could be used for
semanticism and type safety

+++ json.go
@@ -5,7 +5,7 @@ import "encoding/json"
 type auxTable[T freq] struct {
 	N        int       `json:"n"`
 	Freqs    []T       `json:"freqs"`
-	Total    float64   `json:"total"`
+	Total    uint64    `json:"total"`
 	Alphabet *Alphabet `json:"alphabet"`
 }
